+++ libavcodec/arm/mpegvideoencdsp_init_arm.c
@@ -24,8 +24,8 @@
 #include "libavcodec/avcodec.h"
 #include "libavcodec/mpegvideoencdsp.h"
 
-int ff_pix_norm1_armv6(const uint8_t *pix, int line_size);
-int ff_pix_sum_armv6(const uint8_t *pix, int line_size);
+int ff_pix_norm1_armv6(const uint8_t *pix, ptrdiff_t line_size);
+int ff_pix_sum_armv6(const uint8_t *pix, ptrdiff_t line_size);
 
 av_cold void ff_mpegvideoencdsp_init_arm(MpegvideoEncDSPContext *c,
                                          AVCodecContext *avctx)
