+++ libavcodec/x86/mpegvideoencdsp_init.c
@@ -23,9 +23,9 @@
 #include "libavcodec/avcodec.h"
 #include "libavcodec/mpegvideoencdsp.h"
 
-int ff_pix_sum16_sse2(const uint8_t *pix, int line_size);
-int ff_pix_sum16_xop(const uint8_t *pix, int line_size);
-int ff_pix_norm1_sse2(const uint8_t *pix, int line_size);
+int ff_pix_sum16_sse2(const uint8_t *pix, ptrdiff_t line_size);
+int ff_pix_sum16_xop(const uint8_t *pix, ptrdiff_t line_size);
+int ff_pix_norm1_sse2(const uint8_t *pix, ptrdiff_t line_size);
 
 #if HAVE_INLINE_ASM
 
@@ -96,7 +96,7 @@ int ff_pix_norm1_sse2(const uint8_t *pix, int line_size);
 #endif /* HAVE_SSSE3_INLINE */
 
 /* Draw the edges of width 'w' of an image of size width, height */
-static void draw_edges_mmx(uint8_t *buf, int wrap, int width, int height,
+static void draw_edges_mmx(uint8_t *buf, ptrdiff_t wrap, int width, int height,
                            int w, int h, int sides)
 {
     uint8_t *ptr, *last_line;
